reconstructed left zone
